readme and some other cleanup

+++ src/shmup/main.py
@@ -1,11 +1,12 @@
 from __future__ import annotations
 
+import pyglet
+pyglet.options.debug_gl=False
+pyglet.options.vsync=False
+
 import weakref
 
 from math import radians, degrees, floor, ceil, sin, cos
-
-import pyglet
-pyglet.options['debug_gl']=False
 
 from pyglet.gl import glEnable, glDisable, GL_DEPTH_TEST, GL_CULL_FACE, Config
 from pyglet.math import Vec2, Vec3, Mat4, clamp
@@ -38,15 +39,14 @@ class Sound:
             self.player.queue(self.sound)
         self.player.play()        
 
-zap = Sound("zap.wav")
-explosion = [Sound(f"explosion ({n}).wav") for n in range(1,6)]
-oof = Sound("oof.wav")
-dud = Sound("dud.wav")
-
-
+class sounds:
+    zap = Sound("zap.wav")
+    explosion = [Sound(f"explosion ({n}).wav") for n in range(1,6)]
+    oof = Sound("oof.wav")
+    dud = Sound("dud.wav")
 
 r = (-1,0,1)
-s = (0, 1)
+rrr = tuple(product(r,r,r))
 
 import random
 
@@ -67,12 +67,13 @@ class Cube(pyglet.model.Cube, Figure):
     def __init__(self, *a, **ka):
         size = random.choice(self.scales)
         color = random.choice(self.colors)
-        super().__init__(size, size, size, color=color, batch=window.batch)
+        super().__init__(size, size, size, color=color, batch=window.batch, group=window.group)
+        
         self._size = size
 
 class Shot(pyglet.model.Cube):
     def __init__(self, *a, camera:FPSCamera, **ka):
-        super().__init__(.05, .05, 1, (255,255,255,255), batch=window.batch)
+        super().__init__(.05, .05, 1, (255,255,255,255), batch=window.batch, group=window.group)
         self._size = .1
         self.position = Vec3(*camera.position) * Vec3(1,0.75,1)
         self._move = camera._forward*.1
@@ -88,7 +89,7 @@ class Sphere(pyglet.model.Sphere, Figure):
     def __init__(self, *a, **ka):
         size = random.choice(self.scales)
         color = random.choice(self.colors)
-        super().__init__(size, color=color, batch=window.batch)
+        super().__init__(size, color=color, batch=window.batch, group=window.group)
         self._size = size
 
 class Game:
@@ -103,6 +104,7 @@ class Game:
         self.fire_timer = 0
         self.firing = False
 
+    def start(self):
         for _ in range(200):
             item = Cube()
             self.items.append(item)
@@ -132,44 +134,45 @@ class Game:
     def do_shots(self):
         y:Cube|Sphere
         # a=b=0
-        new_shots = []
-        for shot in self.shots:
-            for _ in range(0,5):
-                if shot.position.y<-1:
-                    shot._timer=None
-                    dud.play()
-                    break
-                for a,b in product(r,r):
-                    lx, ly, lz=round(shot.position.x)+a, round(shot.position.y)+b, round(shot.position.z)
-                    if y:=self.space.get((lx, ly, lz), None):  # type: ignore
-                        dist = shot.position.distance(y.pos)
-                        if abs(dist)< y._size-shot._size:
-                            y._vlist.delete()
-                            self.space.pop((lx, ly, lz))
-                            shot._timer=None
-                            random.choice(explosion).play()
-                            break
-                        
-                if shot._timer:
-                    shot.position += shot._move
-                    shot._timer -=1
-                    if not shot._timer:
-                        shot._timer = None
+        
+        if self.shots:
+            new_shots = []
+            for shot in self.shots:
+                for _ in range(0,5):
+                    if shot.position.y<-1:
+                        shot._timer=None
+                        sounds.dud.play()
                         break
-            
-            if shot._timer:
-                shot.matrix = Mat4().translate(shot.position) @ shot.rotation
-                new_shots.append(shot)
-            else:
-                shot._vlist.delete()
-
-        self.shots[:] = new_shots
+                    for a,b,c in rrr:
+                        lx, ly, lz=round(shot.position.x)+a, round(shot.position.y)+b, round(shot.position.z)+c
+                        if y:=self.space.get((lx, ly, lz), None):  # type: ignore
+                            dist = shot.position.distance(y.pos)
+                            if abs(dist)< y._size-shot._size:
+                                y._vlist.delete()
+                                self.space.pop((lx, ly, lz))
+                                shot._timer=None
+                                random.choice(sounds.explosion).play()
+                                break
+                            
+                    if shot._timer:
+                        shot.position += shot._move
+                        shot._timer -=1
+                        if not shot._timer:
+                            shot._timer = None
+                            break
+                
+                if shot._timer:
+                    shot.matrix = Mat4().translate(shot.position) @ shot.rotation
+                    new_shots.append(shot)
+                else:
+                    shot._vlist.delete()
+            self.shots[:] = new_shots
 
         if self.firing and self.fire_timer==0:
             new_shot = Shot(camera=self.camera)
             self.fire_timer = 4
             self.shots.append(new_shot)
-            zap.play()
+            sounds.zap.play()
 
         if self.fire_timer:
             self.fire_timer-=1
@@ -179,12 +182,12 @@ class Game:
         n=0
         while n<10:
             pos += movement
-            for a,b in product(s,s):
-                if i:=self.space.get((round(pos.x)+a, round(pos.y)+b, round(pos.z)),None):
+            for a,b,c in rrr:
+                if i:=self.space.get((round(pos.x)+a, round(pos.y)+b, round(pos.z)+c),None):
                     if i.pos.distance(pos)<i._size:
                         pos -= movement * (n+5)
                         n=10
-                        oof.play()
+                        sounds.oof.play()
                         break
             n+=1
         camera.position = pos
@@ -398,8 +401,11 @@ class FPSCamera:
             self._window.set_exclusive_mouse(False)
             return pyglet.event.EVENT_HANDLED
 
-        if symbol == pyglet.window.key.Z:
+        elif symbol == pyglet.window.key.Z:
             self.game.firing = True
+
+        elif symbol == pyglet.window.key.TAB:
+            window.show_hud = not window.show_hud
             
         
         return False
@@ -450,8 +456,10 @@ class FPSCamera:
 class Window(pyglet.window.Window):
     def __init__(self, *a, **ka):
         config = Config(
-            sample_buffers=1, samples=4, depth_size=16,
-            double_buffer=True
+            sample_buffers=1, samples=8, depth_size=24,
+            double_buffer=True,
+            debug=False,
+            # vsync=False
         )
         
         super().__init__(*a, resizable=True, config=config, **ka)
@@ -464,10 +472,13 @@ class Window(pyglet.window.Window):
         self.view: Mat4
 
         self.view_2d = self.view        
+        self.group = pyglet.graphics.Group()
         self.batch = pyglet.graphics.Batch()
         self.batch2d = pyglet.graphics.Batch()
         self.projection_2d: Mat4
         self.projection_3d: Mat4
+
+        self.show_hud = True
 
         pyglet.font.load("Press Start")
 
@@ -491,25 +502,88 @@ class Window(pyglet.window.Window):
         self.projection=self.projection_2d
         self.batch2d.draw()
 
+        if self.show_hud:
+            fps.draw()
+            fps2.draw()
+            fps3.draw()
+
+
+class FPSDisplay:
+    # update_period = 0.25
+    label: pyglet.text.Label
+
+    def __init__(self, hook, method, color: tuple[int, int, int, int] = (127, 127, 127, 127, ),
+                 samples: int = 240, y=10, label="") -> None:
+        from collections import deque
+        from statistics import mean
+        from time import monotonic
+
+        from pyglet.text import Label
+        self._time = monotonic
+        self._mean = mean
+
+        def wrap(fn):
+            def wrapper(*a, **ka):
+                start = monotonic()
+                result= fn(*a, **ka)
+                self.update(monotonic()-start)
+                return result
+            return wrapper
+        
+        hook2 = wrap(getattr(hook, method))
+        setattr(hook, method, hook2)
+
+        self._label =label
+
+        self.label = Label('', x=10, y=y, font_size=24, weight='bold', color=color)
+        self._delta_times = deque(maxlen=samples)
+        self.counter = 0
+
+    def update(self, t) -> None:
+        """Records a new data point at the current time.
+
+        This method is called automatically when the window buffer is flipped.
+        """
+        self.counter += 1
+        self._delta_times.append(t)
+
+        if self.counter==30:
+            self.counter=0
+            self.label.text = f'{self._label}: {self._mean(self._delta_times):.7f}'
+
+    def draw(self) -> None:
+        """Draw the label."""
+        self.label.draw()
+
+
 window:Window
 camera:FPSCamera
 
 def main():
-    global window, camera
+    global window, camera, fps, fps2, fps3
+
     window = Window()
-    
     camera = FPSCamera(window, position=Vec3(0.0, .5, 5.0))
     game = Game(window,camera)
-    fps = pyglet.window.FPSDisplay(window=window)
+
+    fps = FPSDisplay(game,"do_collisions", label="Collisions")
+    fps2 = FPSDisplay(game, "do_shots", y=36, label="Shots")
+    fps3 = FPSDisplay(window, "on_draw", y=64, label="Draw time")
 
     if controllers := pyglet.input.get_controllers():
         controller = controllers[0]
         controller.open()
         controller.push_handlers(camera)
 
+    import gc
+    gc.freeze()
+
+    game.start()
+
     window.set_visible(True)
     window.flip()
-    pyglet.app.run()
+
+    pyglet.app.run(1/60)
 
 if __name__ == "__main__":
     main()
